Removed CRISPR/Cas9 requirements and made PAM agnostic
GTagHD now only looks at the target sequence, and cuts either between the 3rd and 4th base in the target sequence in the case of antisense guide sequence, and between the 17th and 18th in the case of sense guide sequence. This update changes all functions depending on cut location to use guide sequence orientation to determine cut location.